finalProject/requirements.txt
Adafruit-Blinka==6.4.1
adafruit-circuitpython-apds9960==2.2.5
adafruit-circuitpython-busdevice==5.0.6
adafruit-circuitpython-framebuf==1.4.6
adafruit-circuitpython-motor==3.3.5
adafruit-circuitpython-mpr121==2.1.7
adafruit-circuitpython-mpu6050==1.1.6
adafruit-circuitpython-pca9685==3.3.8
adafruit-circuitpython-register==1.9.14
adafruit-circuitpython-requests==1.12.11
adafruit-circuitpython-seesaw==1.10.0
adafruit-circuitpython-servokit==1.3.5
adafruit-circuitpython-ssd1306==2.11.2
adafruit-circuitpython-typing==1.8.2
Adafruit-GPIO==1.0.3
Adafruit-PlatformDetect==3.4.1
Adafruit-PureIO==1.1.8
Adafruit-SSD1306==1.6.2
pyftdi==0.52.9
pynmea2==1.18.0
pyserial==3.5
pyusb==1.1.1
rpi-ws281x==4.2.6
RPi.GPIO==0.7.1
smbus2==0.4.2
sparkfun-pi-servo-hat==0.9.0
sparkfun-qwiic==1.1.5
sparkfun-qwiic-adxl313==0.0.7
sparkfun-qwiic-as6212==0.0.2
sparkfun-qwiic-bme280==0.9.0
sparkfun-qwiic-button==2.0.1
sparkfun-qwiic-ccs811==0.9.4
sparkfun-qwiic-dual-encoder-reader==0.0.2
sparkfun-qwiic-eeprom==0.0.1
sparkfun-qwiic-gpio==0.0.2
sparkfun-qwiic-i2c==0.9.11
sparkfun-qwiic-icm20948==0.0.1
sparkfun-qwiic-joystick==0.9.0
sparkfun-qwiic-keypad==0.9.0
sparkfun-qwiic-kx13x==1.0.0
sparkfun-qwiic-led-stick==0.0.1
sparkfun-qwiic-max3010x==0.0.2
sparkfun-qwiic-micro-oled==0.10.0
sparkfun-qwiic-oled-base==0.0.2
sparkfun-qwiic-oled-display==0.0.2
sparkfun-qwiic-pca9685==0.9.1
sparkfun-qwiic-pir==0.0.4
sparkfun-qwiic-proximity==0.9.0
sparkfun-qwiic-relay==0.0.2
sparkfun-qwiic-rfid==2.0.0
sparkfun-qwiic-scmd==0.9.1
sparkfun-qwiic-serlcd==0.0.1
sparkfun-qwiic-sgp40==0.0.4
sparkfun-qwiic-soil-moisture-sensor==0.0.2
sparkfun-qwiic-tca9548a==0.9.0
sparkfun-qwiic-titan-gps==0.1.1
sparkfun-qwiic-twist==0.9.0
sparkfun-qwiic-vl53l1x==1.0.1
sparkfun-top-phat-button==0.0.2
sparkfun-ublox-gps==1.1.4
spidev==3.5
sysv-ipc==1.1.0
typing_extensions==4.4.0
